# ayunis-core-frontend/src/shared/api/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/splitter/split": {
      "post": {
        "operationId": "SplitterController_splitText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitTextDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The text has been successfully split into chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "404": {
            "description": "Splitter provider not found"
          }
        },
        "summary": "Split text using the specified provider",
        "tags": [
          "splitter"
        ]
      }
    },
...
    "/runs/stream/{threadId}": {
      "get": {
        "description": "Establishes a server-sent events connection and returns a session ID for sending messages",
        "operationId": "RunsController_connectToStream",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events stream with discriminated response types",
            "content": {
              "text/event-stream": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RunSessionResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/RunMessageResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/RunErrorResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/RunThreadResponseDto"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "session": "#/components/schemas/RunSessionResponseDto",
                      "message": "#/components/schemas/RunMessageResponseDto",
                      "error": "#/components/schemas/RunErrorResponseDto",
                      "thread": "#/components/schemas/RunThreadResponseDto"
                    }
                  }
                },
                "examples": {
                  "session-event": {
                    "summary": "Session establishment event",
                    "value": {
                      "type": "session",
                      "success": true,
                      "threadId": "123e4567-e89b-12d3-a456-426614174000",
                      "timestamp": "2024-01-01T12:00:00.000Z"
                    }
                  },
                  "message-event": {
                    "summary": "Message event",
                    "value": {
                      "type": "message",
                      "message": {
                        "id": "123e4567-e89b-12d3-a456-426614174000",
                        "threadId": "123e4567-e89b-12d3-a456-426614174000",
                        "role": "assistant",
                        "content": [
                          {
                            "type": "text",
                            "text": "Hello!"
                          }
                        ],
                        "createdAt": "2024-01-01T12:00:00.000Z"
                      },
                      "threadId": "123e4567-e89b-12d3-a456-426614174000",
                      "timestamp": "2024-01-01T12:00:00.000Z"
                    }
                  },
                  "error-event": {
                    "summary": "Error event",
                    "value": {
                      "type": "error",
                      "message": "An error occurred while processing your request",
                      "threadId": "123e4567-e89b-12d3-a456-426614174000",
                      "timestamp": "2024-01-01T12:00:00.000Z",
                      "code": "EXECUTION_ERROR"
                    }
                  },
                  "thread-event": {
                    "summary": "Thread update event",
                    "value": {
                      "type": "thread",
                      "threadId": "123e4567-e89b-12d3-a456-426614174000",
                      "updateType": "title_updated",
                      "title": "Discussion about AI and machine learning",
                      "timestamp": "2024-01-01T12:00:00.000Z"
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Type": {
                "description": "Content type for server-sent events",
                "schema": {
                  "type": "string",
                  "example": "text/event-stream"
                }
              },
              "Cache-Control": {
                "description": "Cache control header",
                "schema": {
                  "type": "string",
                  "example": "no-cache"
                }
              },
              "Connection": {
                "description": "Connection type",
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              }
            }
          }
        },
        "summary": "Connect to the run stream and receive a session ID",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/send-message": {
      "post": {
        "description": "Sends a user message to the specified session and triggers AI processing",
        "operationId": "RunsController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent successfully. Response will be streamed to the SSE connection.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Message sent to session"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload or session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Send a message to an active session",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ayunis Core API",
    "description": "The Ayunis Core API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SplitterMetadataDto": {
        "type": "object",
        "properties": {
          "chunkSize": {
            "type": "number",
            "description": "Number of text chunks or lines per segment",
            "example": 1000
          },
          "chunkOverlap": {
            "type": "number",
            "description": "Number of characters to overlap between chunks",
            "example": 200
          },
          "preserveHeader": {
            "type": "boolean",
            "description": "For CSV files: whether to preserve header in each chunk",
            "default": true
          },
          "skipBlankLines": {
            "type": "boolean",
            "description": "For CSV files: whether to skip blank lines",
            "default": true
          },
          "detectQuotes": {
            "type": "boolean",
            "description": "For CSV files: whether to handle quoted fields that span multiple lines",
            "default": true
          },
          "headerRow": {
            "type": "string",
            "description": "For CSV files: custom header row to use instead of first line"
          }
        }
      },
      "SplitTextDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "The text content to split",
            "example": "This is a sample text that needs to be split into chunks..."
          },
          "provider": {
            "type": "string",
            "description": "The splitter provider to use",
            "enum": [
              "recursive",
              "line"
            ],
            "example": "recursive"
          },
          "metadata": {
            "description": "Metadata for customizing the splitting behavior",
            "allOf": [
              {
                "$ref": "#/components/schemas/SplitterMetadataDto"
              }
            ]
          }
        },
        "required": [
          "text",
          "provider"
        ]
      },
      "TextChunkDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "The text content of the chunk"
          },
          "metadata": {
            "type": "object",
            "description": "Metadata about the chunk",
            "example": {
              "index": 0,
              "lineCount": 5,
              "hasHeader": true,
              "startLine": 1,
              "endLine": 5
            }
          }
        },
        "required": [
          "text"
        ]
      },
      "SplitResultDto": {
        "type": "object",
        "properties": {
          "chunks": {
            "description": "The chunks of text produced by the splitting operation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextChunkDto"
            }
          },
          "metadata": {
            "type": "object",
            "description": "Metadata about the split operation",
            "example": {
              "provider": "recursive",
              "chunkSize": 1000,
              "chunkOverlap": 200,
              "totalChunks": 5
            }
          }
        },
        "required": [
          "chunks",
          "metadata"
        ]
      },
...
      "UserMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the message",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID this message belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "role": {
            "type": "string",
            "description": "Role of the message sender",
            "example": "user",
            "enum": [
              "user"
            ]
          },
          "content": {
            "type": "array",
            "description": "Array of text content items for user messages",
            "items": {
              "$ref": "#/components/schemas/TextMessageContentResponseDto"
            }
          }
        },
        "required": [
          "id",
          "threadId",
          "createdAt",
          "role",
          "content"
        ]
      },
...
      "AssistantMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the message",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID this message belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "role": {
            "type": "string",
            "description": "Role of the message sender",
            "example": "assistant",
            "enum": [
              "assistant"
            ]
          },
          "content": {
            "type": "array",
            "description": "Array of content items for assistant messages (text or tool use)",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TextMessageContentResponseDto"
                },
                {
                  "$ref": "#/components/schemas/ToolUseMessageContentResponseDto"
                }
              ]
            }
          }
        },
        "required": [
          "id",
          "threadId",
          "createdAt",
          "role",
          "content"
        ]
      },
      "ToolResultMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the message",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID this message belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "role": {
            "type": "string",
            "description": "Role of the message sender",
            "example": "tool",
            "enum": [
              "tool"
            ]
          },
          "content": {
            "type": "array",
            "description": "Array of tool result content items for tool messages",
            "items": {
              "$ref": "#/components/schemas/ToolResultMessageContentResponseDto"
            }
          }
        },
        "required": [
          "id",
          "threadId",
          "createdAt",
          "role",
          "content"
        ]
      },
...
      "RunSessionResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Response type identifier",
            "example": "session",
            "enum": [
              "session"
            ]
          },
          "success": {
            "type": "boolean",
            "description": "Indicates successful session establishment",
            "example": true
          },
          "streaming": {
            "type": "boolean",
            "description": "Indicates if the session is streaming",
            "example": true
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID for the session",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "timestamp": {
            "type": "string",
            "description": "Session establishment timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          }
        },
        "required": [
          "type",
          "threadId",
          "timestamp"
        ]
      },
      "RunMessageResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Response type identifier",
            "example": "message",
            "enum": [
              "message"
            ]
          },
          "message": {
            "description": "The message data",
            "oneOf": [
              {
                "$ref": "#/components/schemas/UserMessageResponseDto"
              },
              {
                "$ref": "#/components/schemas/AssistantMessageResponseDto"
              },
              {
                "$ref": "#/components/schemas/ToolResultMessageResponseDto"
              },
              {
                "$ref": "#/components/schemas/SystemMessageResponseDto"
              }
            ]
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID for the message",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "timestamp": {
            "type": "string",
            "description": "Message timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          }
        },
        "required": [
          "type",
          "message",
          "threadId",
          "timestamp"
        ]
      },
      "RunErrorResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Response type identifier",
            "example": "error",
            "enum": [
              "error"
            ]
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred while processing your request"
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID where the error occurred",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "timestamp": {
            "type": "string",
            "description": "Error timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          },
          "code": {
            "type": "string",
            "description": "Optional error code for categorization",
            "example": "EXECUTION_ERROR"
          },
          "details": {
            "type": "object",
            "description": "Optional additional error details",
            "example": {
              "originalError": "Network timeout"
            }
          }
        },
        "required": [
          "type",
          "message",
          "threadId",
          "timestamp"
        ]
      },
      "RunThreadResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Response type identifier",
            "example": "thread",
            "enum": [
              "thread"
            ]
          },
          "threadId": {
            "type": "string",
            "description": "Thread ID that was updated",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "updateType": {
            "type": "string",
            "description": "Type of thread update",
            "example": "title_updated",
            "enum": [
              "title_updated"
            ]
          },
          "title": {
            "type": "string",
            "description": "Updated thread title",
            "example": "Discussion about AI and machine learning"
          },
          "timestamp": {
            "type": "string",
            "description": "Thread update timestamp",
            "example": "2024-01-01T12:00:00.000Z"
          }
        },
        "required": [
          "type",
          "threadId",
          "updateType",
          "title",
          "timestamp"
        ]
      },
      "TextInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of input",
            "enum": [
              "text"
            ],
            "example": "text"
          },
          "text": {
            "type": "string",
            "description": "The text content for the inference",
            "example": "What is the weather forecast for New York tomorrow?"
          }
        },
        "required": [
          "type",
          "text"
        ]
      },
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "description": "The thread to use for the inference.",
            "example": "550e8400-e29b-12d3-a456-426614174000"
          },
          "input": {
            "description": "The input to use for the inference",
            "discriminator": {
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextInput"
              },
              {
                "$ref": "#/components/schemas/ToolResultInput"
              }
            ]
          },
          "streaming": {
            "type": "boolean",
            "description": "Enable streaming mode for real-time response updates",
            "example": true,
            "default": false
          }
        },
        "required": [
          "threadId",
          "input"
        ]
      },
...
    }
  }
}